(fix) isolate jmh and unit_test #69
