+++ jraft-example/pom.xml
@@ -40,6 +40,15 @@
             <artifactId>log4j-jcl</artifactId>
             <version>${log4j.version}</version>
         </dependency>
+        <!-- benchmark -->
+        <dependency>
+            <groupId>org.openjdk.jmh</groupId>
+            <artifactId>jmh-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.openjdk.jmh</groupId>
+            <artifactId>jmh-generator-annprocess</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
