+++ jraft-example/src/test/java/com/alipay/sofa/jraft/benchmark/kv/BenchmarkUtil.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.jraft.rhea.benchmark;
+package com.alipay.sofa.jraft.benchmark.kv;
 
 import java.util.concurrent.ThreadLocalRandom;
