+++ jraft-example/src/test/java/com/alipay/sofa/jraft/benchmark/kv/rhea/RheaBenchmarkCluster.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.jraft.rhea.benchmark.rhea;
+package com.alipay.sofa.jraft.benchmark.kv.rhea;
 
 import java.io.File;
 import java.io.IOException;
